+++ windows/WinUtil.h
@@ -575,8 +575,8 @@ public:
 	static int SetupPreviewMenu(CMenu &previewMenu, string extension);
 	static void RunPreviewCommand(unsigned int index, string target);
 	static string formatTime(long rest);
-	static int getImage(const Identity& u);
-	static int getFlagImage(const char* country, bool fullname = false);
+	static u_int8_t getImage(const Identity& u);
+	static u_int8_t getFlagImage(const char* country, bool fullname = false);
 	static string generateStats();
 	static string disableCzChars(string message);
 private:
